src/main/java/sn/sonatel/formation/junit/repository/Lampe.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package sn.sonatel.formation.junit.repository;

/**
 *
 * @author ndiaye028446
 */
public class Lampe {

    String id;
    String emplacement;
    Etat etat;

    public Lampe(String id, String emplacement, Etat etat) {
        this.id = id;
        this.emplacement = emplacement;
        this.etat = etat;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getEmplacement() {
        return emplacement;
    }

    public void setEmplacement(String emplacement) {
        this.emplacement = emplacement;
    }

    public Etat getEtat() {
        return etat;
    }

    public void setEtat(Etat etat) {
        this.etat = etat;
    }

}
